move collection state into separate file
