+++ lib/collection/src/collection_state.rs
@@ -1,0 +1,83 @@
+use crate::{
+    ChannelService, Collection, CollectionConfig, CollectionResult, PeerId, RemoteShard, Shard,
+    ShardId,
+};
+use std::collections::HashMap;
+use std::path::Path;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+pub struct State {
+    pub config: CollectionConfig,
+    pub shard_to_peer: HashMap<ShardId, PeerId>,
+}
+
+impl State {
+    pub async fn apply(
+        self,
+        this_peer_id: PeerId,
+        collection: &mut Collection,
+        collection_path: &Path,
+        channel_service: ChannelService,
+    ) -> CollectionResult<()> {
+        Self::apply_config(self.config, collection).await?;
+        Self::apply_shard_to_peer(
+            self.shard_to_peer,
+            this_peer_id,
+            collection,
+            collection_path,
+            channel_service,
+        )
+        .await
+    }
+
+    async fn apply_config(
+        config: CollectionConfig,
+        collection: &mut Collection,
+    ) -> CollectionResult<()> {
+        log::warn!("Applying only optimizers config snapshot. Other config updates are not yet implemented.");
+        collection
+            .update_optimizer_params(config.optimizer_config)
+            .await
+    }
+
+    async fn apply_shard_to_peer(
+        shard_to_peer: HashMap<ShardId, PeerId>,
+        this_peer_id: PeerId,
+        collection: &mut Collection,
+        collection_path: &Path,
+        channel_service: ChannelService,
+    ) -> CollectionResult<()> {
+        for (shard_id, peer_id) in shard_to_peer {
+            match collection.shards.get(&shard_id) {
+                Some(shard) => {
+                    if shard.peer_id(this_peer_id) != peer_id {
+                        // shard registered on a different peer
+                        log::warn!("Shard movement between peers is not yet implemented. Failed to move shard {shard_id} to peer {peer_id}")
+                    }
+                }
+                None => {
+                    if peer_id == this_peer_id {
+                        // missing local shard
+                        log::warn!("Shard addition is not yet implemented. Failed to add local shard {shard_id}");
+                    } else {
+                        // missing remote shard
+                        let collection_id = collection.id.clone();
+                        let shard_path = crate::create_shard_dir(collection_path, shard_id).await?;
+                        let shard = RemoteShard::init(
+                            shard_id,
+                            collection_id,
+                            peer_id,
+                            shard_path,
+                            channel_service.clone(),
+                        )?;
+                        collection.shards.insert(shard_id, Shard::Remote(shard));
+                        collection.ring.add(shard_id);
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+}
